feat(nvim): Add an org-mode keybind to follow links more loosely
Looks for an org-mode link on or below the current Treesitter node and follows it, rather than requiring the cursor to be on the link node.

NOTE: This assumes only one link is in the tree below the cursor node. If multiple link nodes are present, the first one found will be followed.